requirements.txt
pyTelegramBotAPI==4.14.0
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30
datetime
json5==0.9.14
bcrypt==4.1.2
PyJWT==2.10.1

